@@ -7,66 +7,46 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "test": "ng test",
+    "test": "jest",
+    "test:watch": "jest --watch",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "cypress open",
+    "e2e:headless": "cypress run"
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~12.0.1",
-    "@angular/core": "~12.0.1",
-    "@angular/forms": "~12.0.1",
-    "@angular/platform-browser": "~12.0.1",
-    "@angular/platform-browser-dynamic": "~12.0.1",
-    "@angular/router": "~12.0.1",
-    "@capacitor/android": "3.0.1",
-    "@capacitor/app": "1.0.1",
-    "@capacitor/core": "3.0.1",
-    "@capacitor/haptics": "1.0.1",
-    "@capacitor/keyboard": "1.0.1",
-    "@capacitor/status-bar": "1.0.1",
-    "@capacitor/storage": "^1.0.1",
-    "@ionic-native/core": "^5.33.1",
-    "@ionic-native/geolocation": "^5.33.1",
-    "@ionic/angular": "^5.5.2",
-    "cordova-plugin-geolocation": "^4.1.0",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.11.4"
+    "@angular/animations": "^19.2.0",
+    "@angular/common": "^19.2.0",
+    "@angular/compiler": "^19.2.0",
+    "@angular/core": "^19.2.0",
+    "@angular/forms": "^19.2.0",
+    "@angular/platform-browser": "^19.2.0",
+    "@angular/platform-browser-dynamic": "^19.2.0",
+    "@angular/router": "^19.2.0",
+    "@capacitor/android": "^6.0.0",
+    "@capacitor/core": "^6.0.0",
+    "@capacitor/geolocation": "^6.0.0",
+    "@capacitor/preferences": "^6.0.0",
+    "@ionic/angular": "^8.0.0",
+    "ionicons": "^7.0.0",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.8.0",
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.0.1",
-    "@angular-eslint/builder": "~12.0.0",
-    "@angular-eslint/eslint-plugin": "~12.0.0",
-    "@angular-eslint/eslint-plugin-template": "~12.0.0",
-    "@angular-eslint/template-parser": "~12.0.0",
-    "@angular/cli": "~12.0.1",
-    "@angular/compiler": "~12.0.1",
-    "@angular/compiler-cli": "~12.0.1",
-    "@angular/language-service": "~12.0.1",
-    "@capacitor/cli": "3.0.1",
-    "@ionic/angular-toolkit": "^4.0.0",
-    "@ionic/lab": "3.2.10",
-    "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "4.16.1",
-    "@typescript-eslint/parser": "4.16.1",
-    "eslint": "^7.6.0",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-jsdoc": "30.7.6",
-    "eslint-plugin-prefer-arrow": "1.2.2",
-    "jasmine-core": "~3.7.1",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.3.2",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "typescript": "~4.2.4"
+    "@angular-devkit/build-angular": "^19.2.0",
+    "@angular/cli": "^19.2.0",
+    "@angular/compiler-cli": "^19.2.0",
+    "@capacitor/cli": "^6.0.0",
+    "@types/jest": "^29.5.0",
+    "@types/node": "^22.0.0",
+    "angular-eslint": "19.8.1",
+    "cypress": "^14.0.0",
+    "eslint": "^9.28.0",
+    "jest": "^29.7.0",
+    "jest-preset-angular": "^14.4.0",
+    "typescript": "~5.7.0",
+    "typescript-eslint": "8.33.1"
   },
   "description": "An Ionic project"
 }
