@@ -1,35 +1,65 @@
+/* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-inferrable-types */
+import { Component, inject, CUSTOM_ELEMENTS_SCHEMA, OnInit } from '@angular/core';
+import { NgIf } from '@angular/common';
 import { HttpResponse } from '@angular/common/http';
-import { Component, OnInit } from '@angular/core';
+import {
+  IonHeader,
+  IonToolbar,
+  IonTitle,
+  IonContent,
+  IonMenuButton,
+  IonCard,
+  IonCardHeader,
+  IonCardTitle,
+  IonCardContent,
+  IonButton,
+  IonProgressBar,
+  IonImg
+} from '@ionic/angular/standalone';
 import { Cat } from '../models/cat';
 import { CatService } from '../services/cat.service';
 
 @Component({
   selector: 'app-cat',
+  standalone: true,
+  imports: [
+    NgIf,
+    IonHeader,
+    IonToolbar,
+    IonTitle,
+    IonContent,
+    IonMenuButton,
+    IonCard,
+    IonCardHeader,
+    IonCardTitle,
+    IonCardContent,
+    IonButton,
+    IonProgressBar,
+    IonImg
+  ],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA],
   templateUrl: './cat.component.html',
-  styleUrls: ['./cat.component.scss'],
+  styleUrls: ['./cat.component.scss']
 })
 export class CatComponent implements OnInit {
-
-  httpOk: boolean;
+  private catApi = inject(CatService);
+  httpOk: boolean = false;
   cat!: Cat;
 
-  constructor(private catApi: CatService) { }
-
   ngOnInit() {
-    this.catApi.getCat().subscribe((resp: HttpResponse<Cat[]>)=>{
+    this.catApi.getCat().subscribe((resp: HttpResponse<Cat[]>) => {
       console.log(resp);
       this.httpOk = resp.ok;
-      this.cat = resp.body[0];
+      this.cat = resp.body![0];
     });
   }
 
-  refreshCat(){
+  refreshCat() {
     console.log('un autre !');
-    this.catApi.getCat().subscribe((resp: HttpResponse<Cat[]>)=>{
+    this.catApi.getCat().subscribe((resp: HttpResponse<Cat[]>) => {
       console.log(resp);
       this.httpOk = resp.ok;
-      this.cat = resp.body[0];
+      this.cat = resp.body![0];
     });
   }
-
 }
