@@ -1,19 +1,56 @@
-import { Component, OnInit } from '@angular/core';
+/* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-inferrable-types */
+import { Component, inject, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
+import { NgIf, NgFor } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+import {
+  IonHeader,
+  IonToolbar,
+  IonTitle,
+  IonContent,
+  IonMenuButton,
+  IonItem,
+  IonLabel,
+  IonInput,
+  IonButton,
+  IonIcon
+} from '@ionic/angular/standalone';
 import { MeteoService } from '../services/meteo.service';
-import { Geolocation } from '@ionic-native/geolocation/ngx';
+import { MeteoCardComponent } from './meteo-card/meteo-card.component';
+import { Geolocation } from '@capacitor/geolocation';
+import { addIcons } from 'ionicons';
+import { pin } from 'ionicons/icons';
+
 @Component({
   selector: 'app-meteo',
+  standalone: true,
+  imports: [
+    NgIf,
+    NgFor,
+    FormsModule,
+    IonHeader,
+    IonToolbar,
+    IonTitle,
+    IonContent,
+    IonMenuButton,
+    IonItem,
+    IonLabel,
+    IonInput,
+    IonButton,
+    IonIcon,
+    MeteoCardComponent
+  ],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA],
   templateUrl: './meteo.component.html',
-  styleUrls: ['./meteo.component.scss'],
+  styleUrls: ['./meteo.component.scss']
 })
-export class MeteoComponent implements OnInit {
-
-  city: string;
-
+export class MeteoComponent {
+  private meteoService = inject(MeteoService);
   weather: any;
-  constructor(private meteoService: MeteoService, private geolocation: Geolocation) { }
+  city: string = '';
 
-  ngOnInit() { }
+  constructor() {
+    addIcons({ pin });
+  }
 
   getWeatherByCity() {
     this.meteoService.getWeatherByCity(this.city)
@@ -29,11 +66,9 @@ export class MeteoComponent implements OnInit {
   }
 
   async getWeatherByLocation() {
-    this.geolocation.getCurrentPosition().then((geoloc: GeolocationPosition) => {
-      this.meteoService.getWeatherByLocation(geoloc.coords).subscribe(
-        resp => this.weather = resp
-      );
-    });
+    const geoloc = await Geolocation.getCurrentPosition();
+    this.meteoService.getWeatherByLocation(geoloc.coords).subscribe(
+      resp => this.weather = resp
+    );
   }
-
 }
