@@ -1,4 +1,4 @@
-import { Injectable } from '@angular/core';
+import { Injectable, inject } from '@angular/core';
 import { HttpClient, HttpResponse } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { Cat } from '../models/cat';
@@ -8,13 +8,9 @@ import { environment } from '../../environments/environment';
   providedIn: 'root'
 })
 export class CatService {
+  private http = inject(HttpClient);
 
-  catObs: Observable<Cat[]>;
-
-  constructor(private http: HttpClient) { }
-
-  getCat(): Observable<HttpResponse<Cat[]>>{
-    return this.http.get<Cat[]>(environment.catUrl, {observe: 'response'});
+  getCat(): Observable<HttpResponse<Cat[]>> {
+    return this.http.get<Cat[]>(environment.catUrl, { observe: 'response' });
   }
-
 }
